# server/openapi.yml
openapi: 3.0.0
info:
  title: Fulcrum API
  version: 1.0.0
  description: Fulcrum is a support ticket management and rating system
  license:
    name: MIT
  contact:
    name: Amol HATWAR
    email: 21f1000451@ds.study.iitm.ac.in
    url: http://onlinedegree.iitm.ac.in/
servers:
  - url: http://localhost:8000/api
    description: Fulcrum Development Server
components:
  securitySchemes:
    APIKeyAuth:
      type: apiKey
      in: header
      name: Authentication-Token
  responses:
    MissingData:
      description: Required field missing
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
              description:
                type: object
              name:
                type: string
    Unauthenticated:
      description: Unauthenticated Operation
      content:
        application/json:
          schema:
              type: object
              properties:
                meta:
                  type: object
                  properties:
                    code:
                      type: integer
                response:
                  type: object
                  properties:
                    errors:
                      type: array
                      items:
                        type: string
paths:
  /signup:
    post:
      summary: New user signup as student
      operationId: user.signup
      tags:
        - users
      responses:
        422:
          description: Signup with missing email
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    code:
                      type: integer
                    description:
                      type: string
                    name:
                      type: string
        409:
          description: User with given email already exists
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    code:
                      type: integer
                    description:
                      type: string
                    name:
                      type: string
        201:
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    message:
                      type: string
                    status:
                      type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - first_name
                - last_name
                - roll_number
                - email
                - password
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                roll_number:
                  type: string
                email:
                  type: string
                password:
                  type: string
            examples:
              good:
                value:
                  first_name: Amol
                  last_name: Hatwar
                  email: amol@hatwar.org
                  roll_number: 21f1000451
                  password: ThisIsATest
              duplicate:
                value:
                  first_name: Amol
                  last_name: Hatwar
                  email: amol@hatwar.org
                  roll_number: 21f1000451
                  password: ThisIsATest
              missing_email:
                value:
                  first_name: Amol
                  last_name: Hatwar
                  roll_number: 21f1000451
                  password: ThisIsATest
  /signin:
    post:
      summary: User Signin
      operationId: user.signin
      tags:
        - users
      responses:
        201:
          description: User session created
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    auth_token:
                      type: string
                    message:
                      type: string
                    status:
                      type: string
        401:
          description: Bad email or password
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    code:
                      type: integer
                    description:
                      type: string
                    name:
                      type: string
        422:
          $ref: '#/components/responses/MissingData'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
            examples:
              good:
                value:
                  email: amol@hatwar.org
                  password: ThisIsATest
              bad:
                value:
                  email: amol@hatwar.org
                  password: ThisIsAlsoTest
              missing_email:
                value:
                  email: amol@hatwar.org
                  password: ''
  /profile:
    get:
      summary: View User Profile
      operationId: user.profile
      tags:
        - users
      security:
        - APIKeyAuth: []
      responses:
        200:
          description: User profile shown
          content:
            application/json:
              schema:
                  type: object
                  properties:
                    id:
                      type: integer
                    first_name:
                      type: string
                    last_name:
                      type: string
                    roll_number:
                      type: string
                    email:
                      type: string
        401:
          $ref: '#/components/responses/Unauthenticated'
  /signout:
    post:
      summary: Signout User
      operationId: user.signout
      tags:
        - users
      security:
        - APIKeyAuth: []
      responses:
        204:
          description: User signed out
          content:
            application/json:
              schema:
                  type: object
        401:
          $ref: '#/components/responses/Unauthenticated'
  /profile/update:
    put:
      summary: User updating self
      operationId: user.profile.update
      tags:
        - users
      security:
        - APIKeyAuth: []
      responses:
        422:
          $ref: '#/components/responses/MissingData'
        401:
          $ref: '#/components/responses/Unauthenticated'
        204:
          description: Successful Update
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    message:
                      type: string
                    status:
                      type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - first_name
                - last_name
                - email
                - password
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                roll_number:
                  type: string
                email:
                  type: string
                password:
                  type: string
            examples:
              good:
                value:
                  first_name: Amol
                  last_name: Hatwar
                  email: 21f1000451@ds.study.iitm.ac.in
                  password: ThisIsATest
              with_roll:
                value:
                  first_name: Amol
                  last_name: Hatwar
                  email: amol@hatwar.org
                  roll_number: 21f1000451
                  password: ThisIsATest
              missing_email:
                value:
                  first_name: Amol
                  last_name: Hatwar
                  password: ThisIsATest